requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydicom==2.4.2
numpy==1.26.4
opencv-python-headless==4.8.1.78
python-multipart==0.0.6
pillow==10.0.0
python-dotenv==1.0.1